requirements_2.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1
django-cors-headers==4.3.1
whitenoise==6.2.0
gunicorn==21.2.0
dj-database-url==0.5.0
tzdata==2025.2